+++ config/config.go
@@ -7,10 +7,10 @@ var (
 	repoURL = "https://github.com/sfborg/sfga"
 
 	// tag of the sfga repo to get correct schema version.
-	repoTag = "v0.3.24"
+	repoTag = "v0.3.25"
 
 	// schemaHash is the sha256 sum of the correponding schema version.
-	schemaHash = "b1db9df2e759f"
+	schemaHash = "bdab0c86c9fa"
 )
 
 type Config struct {