+++ io/sfgaio/coldp-author.go
@@ -19,14 +19,15 @@ func (s *sfgaio) InsertAuthors(data []coldp.Author) error {
 	}()
 
 	stmt, err := tx.Prepare(`
-	INSERT INTO author
-		(
-			id, source_id, alternative_id, given, family, suffix,
-	  	abbreviation_botany, alternative_names, sex_id, country, birth,
-	  	birth_place, death, affiliation, interest, reference_id, link,
-	  	remarks, modified, modified_by
-		)
-	VALUES (?,?,?,?,?,?, ?,?,?,?,?, ?,?,?,?,?,?, ?,?,?)
+  INSERT INTO author
+    (
+    col__id, col__source_id, col__alternative_id, col__given, col__family,
+    col__suffix, col__abbreviation_botany, col__alternative_names, col__sex_id,
+    col__country, col__birth, col__birth_place, col__death, col__affiliation,
+    col__interest, col__reference_id, col__link, col__remarks, col__modified,
+    col__modified_by
+    )
+  VALUES (?,?,?,?,?, ?,?,?,?, ?,?,?,?,?, ?,?,?,?,?, ?)
 `)
 	if err != nil {
 		return err
@@ -35,10 +36,11 @@ func (s *sfgaio) InsertAuthors(data []coldp.Author) error {
 
 	for _, n := range data {
 		_, err = stmt.Exec(
-			n.ID, n.SourceID, n.AlternativeID, n.Given, n.Family, n.Suffix,
-			n.AbbreviationBotany, n.AlternativeNames, n.Sex, n.Country,
-			n.BirthPlace, n.Death, n.Affiliation, n.Interest, n.ReferenceID,
-			n.Link, n.Remarks, n.Modified, n.ModifiedBy,
+			n.ID, n.SourceID, n.AlternativeID, n.Given, n.Family,
+			n.Suffix, n.AbbreviationBotany, n.AlternativeNames, n.Sex,
+			n.Country, n.Birth, n.BirthPlace, n.Death, n.Affiliation,
+			n.Interest, n.ReferenceID, n.Link, n.Remarks, n.Modified,
+			n.ModifiedBy,
 		)
 		if err != nil {
 			return err