+++ io/sfgaio/coldp-distr.go
@@ -20,11 +20,12 @@ func (s *sfgaio) InsertDistributions(data []coldp.Distribution) error {
 
 	stmt, err := tx.Prepare(`
 	INSERT INTO distribution
-		(
-			taxon_id, source_id, area, area_id, gazetteer_id,
-			status_id, reference_id, remarks, modified, modified_by
-		)
-	VALUES (?,?,?,?,?, ?,?,?,?,?)
+    (
+    col__taxon_id, col__source_id, col__area, col__area_id,
+    col__gazetteer_id, col__status_id, col__reference_id, col__remarks,
+    col__modified, col__modified_by
+    )
+	VALUES (?,?,?,?, ?,?,?,?, ?,?)
 `)
 	if err != nil {
 		return err