+++ io/sfgaio/coldp-synonym.go
@@ -19,13 +19,13 @@ func (s *sfgaio) InsertSynonyms(data []coldp.Synonym) error {
 	}()
 
 	stmt, err := tx.Prepare(`
-	INSERT INTO synonym
-	(
-		id, taxon_id, source_id, name_id, name_phrase,
-		according_to_id, status_id, reference_id,
-		link, remarks, modified, modified_by
-	)
-	VALUES (?,?,?,?,?, ?,?,?, ?,?,?,?)
+  INSERT INTO synonym
+  (
+    col__id, col__taxon_id, col__source_id, col__name_id, col__name_phrase,
+    col__according_to_id, col__status_id, col__reference_id,
+    col__link, col__remarks, col__modified, col__modified_by
+  )
+  VALUES (?,?,?,?,?, ?,?,?, ?,?,?,?)
 `)
 	if err != nil {
 		return err