+++ io/sfgaio/coldp-taxon.go
@@ -19,22 +19,24 @@ func (s *sfgaio) InsertTaxa(data []coldp.Taxon) error {
 	}()
 
 	stmt, err := tx.Prepare(`
-	INSERT INTO taxon
-		(
-		id, alternative_id, source_id, parent_id, ordinal, branch_length,
-		name_id, name_phrase, according_to_id, according_to_page,
-		according_to_page_link, scrutinizer, scrutinizer_id,
-		scrutinizer_date, status_id, reference_id, extinct,
-		temporal_range_start_id, temporal_range_end_id,
-		environment_id, species, section, subgenus, genus, subtribe,
-		tribe, subfamily, family, superfamily, suborder, "order",
-		subclass, class, subphylum, phylum, kingdom,
-		link, remarks, modified, modified_by
-		)
-	VALUES (
-		?,?,?,?,?,?, ?,?,?,?, ?,?,?, ?,?,?,?, ?,?, ?,?,?,?,?,?,
-		?,?,?,?,?,?, ?,?,?,?,?, ?,?,?,?
-		)
+  INSERT INTO taxon
+    (
+    col__id, col__alternative_id, col__source_id, col__parent_id,
+    col__ordinal, col__branch_length, col__name_id, col__name_phrase,
+    col__according_to_id, col__according_to_page,
+    col__according_to_page_link, col__scrutinizer, col__scrutinizer_id,
+    col__scrutinizer_date, col__status_id, col__reference_id, col__extinct,
+    col__temporal_range_start_id, col__temporal_range_end_id,
+    col__environment_id, col__species, col__section, col__subgenus,
+    col__genus, col__subtribe, col__tribe, col__subfamily, col__family,
+    col__superfamily, col__suborder, col__order, col__subclass, col__class,
+    col__subphylum, col__phylum, col__kingdom, col__link, col__remarks,
+    col__modified, col__modified_by
+    )
+  VALUES (
+    ?,?,?,?,?,?, ?,?,?,?, ?,?,?, ?,?,?,?, ?,?, ?,?,?,?,?,?,
+    ?,?,?,?,?,?, ?,?,?,?,?, ?,?,?,?
+    )
 `)
 	if err != nil {
 		return err